requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
sendgrid
python-dotenv